Gain coin and increase kills in enemy death
