+++ Assets/Scripts/PatrolAgent.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Enemy;
 using UnityEngine;
 
 public class PatrolAgent : MonoBehaviour
